+++ MyfirstMavenProjectIT/config.properties
@@ -1,6 +1,7 @@
 driverdir=C:\\Users\\junaid\\Desktop\\Selenium\\
 browser=chrome
 systemUnderTest=http://demo.guru99.com/telecom/index.html
+secondPage=http://demo.guru99.com/telecom/addcustomer.php
 csvdir=C:\\tmp\\creditCard.csv
 csvlogin=C:\\tmp\\login.csv
 driver=C:\\Users\\junaid\\Desktop\\Selenium\\
