+++ MyfirstMavenProjectIT/src/test/java/frameworkclasses/BasePage.java
@@ -6,10 +6,13 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.Properties;
 
+import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.edge.EdgeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.annotations.AfterSuite;
 import org.testng.annotations.Parameters;
 
@@ -84,6 +87,23 @@ public class BasePage {
 	}
 	
 
-
+	public void waitForElement(int elementWait,By pLocator) {
+		WebDriverWait wait = new WebDriverWait(BasePage.driver,elementWait);
+    	wait.until(ExpectedConditions.visibilityOfElementLocated(pLocator));
+    	
+	}
+	
+	public void waitForClick(int elementWait,By pLocator) {
+		WebDriverWait wait = new WebDriverWait(BasePage.driver,elementWait);
+    	wait.until(ExpectedConditions.elementToBeClickable(pLocator));
+    	
+	}
+	
+	public String getheader(By pLocator) {
+		waitForElement(10,pLocator);
+		String h1 = driver.findElement(pLocator).getText();
+		
+		return h1;
+	}
 }
